Moved compass to components dir and updated its parent's imports
