+++ src/components/phasetimer.js
@@ -4,10 +4,9 @@ import "../components/bootstrap.css"
 import { Row, ButtonGroup } from 'react-bootstrap';
 import { Col } from 'react-bootstrap';
 import Timer from 'react-compound-timer';
-import Compass from '../pages/Compass';
+import Compass from './Compass';
 import { log_list } from '../dummyData';
 import { render } from 'react-testing-library';
-
 
  const generateList=(phase)=> {
         if (Compass.currentPhase === phase) {
